Get sandboxes older than 24h
In order to work on cost reports, make sure sandboxes older than 24 are
used when getting a sandbox from the pool.

see https://issues.redhat.com/browse/GPTEINFRA-558

I tried to do LIFO but with dynamodb being a NOSQL database, it's
difficult.

It's not possible to create an index on a column that is a boolean, like `available`.

In order to do that, i would have to create a new column 'isAvailable',
as a string with content `"x"` and absent if not available. It would be
redundant with the current `available` BOOL column.

Then i would need to create an sparse index in dynamoDB with that primary key
`isAvailable`, with the sort key being `aws:rep:updatetime` or a new
manual field "last updated". That sparse index would then be used when
scanning or Quering the DB in order to get the result in a specific
order (sorted by last used for example).

This would imply way more work and touching other parts of the code, not
just the part that gets a new sandbox from the DB.

I suggest with start with the 24h condition when selecting and see if it's enough.